backend/requirements.txt
Flask>=3.1.2
python-dotenv>=1.1.1
psycopg2-binary>=2.9.10
PyJWT>=2.10.1